client/requirements.txt
Flask-SocketIO==5.3.6
python-socketio[client]==5.10.0